Introduce RDWR and RDWR* signal. Move generation of signal close to 8085.
The original schematic created an enable signal for the LCD out of
~RD and ~RW right at the LCD connector; then used the output to create
a buffered version of it called (a)* (or Ⓐ* if this unicode works) which
was somewhat meaningless and thus confusing.

Now, let's create the first signal directly near the M20 buffer,
and call it RDWR. Call the buffered version RDWR* and use it in the
system bus and RAM decoding logic.